add two hooks separeted
